Refactor the code by moving files into packages.
