+++ src/components/Table.js
@@ -1,5 +1,5 @@
 import React from "react";
-import './css/Table.css'
+import '../css/Table.css'
 
 class Table extends React.Component {
     constructor( props ) {
