+++ src/resource/flagsImageDay.png
